pyproject.toml
[project]
name = "14_Biolit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "polars",
    "pandas",
    "sqlalchemy",
    "psycopg2-binary",
    "requests",
    "geopandas",
    "shapely",
    "pyarrow",
    "python-dotenv",
    "structlog",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pytest>=9.0.2",
    "requests>=2.32.3",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.44",
    "prefect>=3.6.22",
    "requests>=2.32.3",
    "shapely>=2.1.2",
    "structlog>=25.5.0",
    "tqdm>=4.67.3",
    "transformers>=4.45",
    "torch>=2.11.0",
    "roboflow>=1.2.16",
    "autodistill>=0.1.29",
    "autodistill-grounding-dino>=0.1.4",
    "autodistill-yolov8>=0.1.4",
    "ultralytics",
    "supervision>=0.27.0.post2",
    "scikit-learn>=1.8.0",
    "python-dotenv>=1.2.2",
    "huggingface-hub>=0.36.2",
    "ultralytics>=8.0.81",
    "minio>=7.2.20",
    "label-studio-sdk>=2.0.13",
    "boto3>=1.42.89",
]

[tool.pytest.ini_options]
pythonpath = ["."]
